requirements.txt
argon2-cffi==20.1.0
asgiref==3.3.1
astroid==2.4.2
cffi==1.14.4
Django==3.1.5
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
PyJWT==2.0.1
pytz==2020.5
sqlparse==0.4.1
gunicorn==20.0.4
django-cors-headers==3.6.0
argon2-cffi==20.1.0
pycparser==2.20